refactor: mengubah underscore pada value enum dengan spasi dan mengubah model kurikulum & feat: menambahkan resource calon santri
Mengubah semua underscore pada value enum dengan spasi dan mengubah model kurikulum
Menambahkan resource calon santri

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->string('nomor_telepon', 16)->unique();
             $table->string('email', 96)->unique();
             $table->string('kelas');
-            $table->unsignedSmallInteger('angkatan_pondok');
+            $table->unsignedInteger('angkatan_pondok');
             $table->string('status_pondok');
             $table->date('tanggal_lulus_pondok')->nullable();
             $table->string('password');
